Wrote these changes locally over 2 weeks ago. I do not remember what they are

+++ framework/src/parser/ast/expression/MemberAccess.cpp
@@ -101,7 +101,7 @@ namespace parser {
         mClass->typeCheck(diag, exit);
 
         if (mClass->getType()->isArrayView() && mId == "length") {
-            mType = Type::Get("int");
+            mType = Type::Get("long");
             return;
         }
